starboard: Permit DRM ticket re-take for provisioning retries (#8279)
Calling SessionUpdateRequest::ReleaseTicket() multiple times during DRM
provisioning retries led to a crash due to an SB_CHECK_NE assertion, but
which can happen on some devices.

This PR renames ReleaseTicket() to TakeTicket() and removes the
assertion. TakeTicket() now returns kSbDrmTicketInvalid on subsequent
calls for the same request, signaling StarboardCdm to treat it as a
spontaneous message for an existing session. 

This ensures robust handling of multi-stage DRM provisioning retries
without crashing.

Bug: 79941850