Enable sanitizers for linux builds (#6350)
This enables asan by default for linux debug/devel builds, and allows
asan to be enabled for Evergreen linux builds (which necessitates the
python changes for symbolization scripts). We don't enable asan by
default on Evergreen as it adds to the list of leaking symbols.

Note there are a couple of changes here to resolve leaking code paths
that seem to be newly brought in because of asan.

Bug: 423906285
Bug: 467135717