Remove symlink files (#8337)
This assumes some tech debt, since it affects `//content`. However, the
cost of engineering a copy action versus this is too high, as every day
the browsertests are not running in CI, the greater chance that the
browsertests are broken by any CL. This has already happened once. So,
the plan of action is to land this, get browsertests running in CI, then
address this tech debt as tracked in b/467370154.

Test: `gn clean out/android-x86_qa && time autoninja -C
out/android-x86_qa cobalt_browsertests && export
ANDROID_SERIAL=localhost:27348 &&
out/android-x86_qa/bin/run_cobalt_browsertests -v -f
*NavigationBrowserTest.SubFrameJsNavigationInitiator*`
Bug: 432313695