tvOS: Plug //starboard/tvos/shared files into the build system
(Note that the build will only work once #8350 and #8463 land)

Now that a lot of tvOS patches have been upstreamed and the code is almost
building as-is (although not fully working at runtime), it is time to plug
most files into GN.

Once the PRs above land, it should be possible to build targets like
`cobalt`, `starboard_unittests_wrapper` and `nplb`.

Not all files in `//starboard/tvos/shared` have been added to `BUILD.gn`.
Some of them either need more work to build but were not blocking the build
and basic tests, or just need to be removed altogether in the future.

Bug: 433180638
Bug: 456692793