android: Dismiss network error dialog once connected to Internet (#8244)
This PR introduces a `NetworkChangeNotifier` to the
`CobaltConnectivityDetector.java` that watches for changes in network
connectivity, and in this case specifically, for reconnection to an
internet source. If a device goes from being unconnected to connected,
another `activeNetworkCheck()` is performed and dismisses the dialog if
it exists if the internet connection is valid. As observed in
b/463699732#comment13. This PR also introduces a `WebContentsObserver`
to `CobaltActivity.java` to determine whether a hard reload is required
on reconnection to the internet (like on startup), or if the dialog
should just dismiss without a reload (like when a video is playing).

Bug: 464033064
Bug: 463699732
(cherry picked from commit 95d33e179786afa0fc2f4318482dc33bae326007)