evergreen: Add EG version to the update check request (#8378)
Include the current Evergreen version string in the extra parameters
sent with the update check request. Evergreen version is often used to
target devices for enabling/disabling updates. In the past, this field
is parsed from the ua string, but sending it directly is more reliable.
Also changed the inserting method to `emplace()` as the preferred style.

Issue: 431862767