Cobalt: Disable libvpx
This change disables the libvpx library, including VP8 and VP9 codec
support, within WebRTC for Cobalt builds. This helps to reduce
the overall binary size of the Cobalt application by removing unused
video codec implementations. Related build targets and code paths in
Blink and WebRTC are now conditionally compiled based on these new
build flags to ensure that the disabled codecs are not included.

The binary size reduction is ~ 700KB.
Bug: 421171908