starboard: Avoid calling Join() in DrmSystem when thread is not started
Previously, starboard::Thread required Join() to be called, even if
Start() had not been called. However, calling Join() without Start()
causes a crash because pthread_join() fails.

This CL removes the requirement to call Join() in the starboard::Thread
destructor and updates DrmSystem to skip joining when app provisioning
is enabled, as the thread is not started in that mode.

Bug: 469772944