cobalt/test: Add race condition tests for navigation commits
Introduces new test fixtures (`UndoCommitNavigationBrowserTest`, `NavigationQueueingBrowserTest`) to specifically test race conditions that can occur when a new navigation begins while a previous one is in a pending commit state.

This is one of several changes to navigation_browsertests as part of the effort to enable browser tests on the main branch for Cobalt.
Note: This is branched on top of #8443 as part of a chain.

Bug: 459831767