cobalt/test: Add Remaining missing tests - Part 4
Ports over a number of missing browser tests from the content_browsertests target and modifies various API surfaces from the M114-M138 rebase. This is the fifth and final in a multi-part effort.

This is one of several changes to navigation_browsertests as part of the effort to enable browser tests on the main branch for Cobalt.
Note: This is branched on top of #8450 as part of a chain.

Bug: 459831767