tvOS: Support App Intents (Siri and dictation) (#8424)
Patch by @jkim-julie with some suggestions by me.

Bring back this feature from C25 but integrate it with the current
Chrobalt architecture:

* Do not use a completely separate Xcode project to build the Swift code
that integrates with the App Intents framework. Instead, use Chromium's
existing infrastructure to build Swift code with GN.

* There is nothing Starboard-specific in the code since the UI bits are
handled in Cobalt Shell, so the code now lives in `//cobalt`. There is
no `SBProcessAppIntent()`, so the Swift code communicates directly with
the Deep Link code via the newly-added DeepLinkSupportTvos interface.

The behavior should match C25's and the official app's, including bugs
and limitations that are described in the bugs below.

Bug: 461350746
Bug: 465023735

---------

Co-authored-by: Julie Jeongeun Kim <jkim@igalia.com>