cobalt/test:  Add tests for credentialless iframes and sandbox warnings (#8445)
Adds new browser tests to verify the behavior of security-related iframe
attributes:
- Ensures cookies are correctly handled for iframes with the
`credentialless` attribute.
- Verifies that a console warning is emitted for iframes with an
escapable sandbox policy (`allow-scripts` and `allow-same-origin`).

This is one of several changes to navigation_browsertests as part of the
effort to enable browser tests on the main branch for Cobalt.
Note: This is branched on top of #8444 as part of a chain.

Bug: 459831767