cobalt/test: Fix and enable disabled browser tests (#8446)
Enables a number of previously disabled navigation browser tests by
addressing underlying flakiness or bugs. This improves test coverage and
suite stability for navigation.

This is one of several changes to navigation_browsertests as part of the
effort to enable browser tests on the main branch for Cobalt.
Note: This is branched on top of #8445 as part of a chain.

Bug: 459831767