cobalt/test: Add Remaining missing tests - Part 1 (#8448)
Port over a number of missing browser tests related to navigation from
the content_browsertests target. This updates various API surfaces
to align with the M114-M138 rebase. This is part of an ongoing effort
to enable browser tests on the main branch for Cobalt.

Note: This is branched on top of #8446 as part of a chain.

Bug: 459831767