cobalt: Make CoAT app always debuggable/profileable
CoAT builds are intended for testing and debugging. Since this manifest
is not used for production releases, these flags are now hardcoded to
always be enabled.

This simplifies development by removing the need to track which build
type is currently in use when attempting to attach a debugger or
run a profiling tool.

Bug: 462494527