h5vcc_accessibility: Split port-specific code into separate files (#8368)
This will make it easier to add a tvOS implementation in a follow-up PR
and also makes the code cleaner in general.

The idea is for `H5vccAccessibilityImpl` to act as a
platform-independent class whose task is to interface with Mojo and
communicate with the platform-specific code in the
`PlatformTextToSpeechHelper` class. This pattern is also found in other
parts of Chromium, like `//services/device/generic_sensor`.

`PlatformTextToSpeechHelper` is subclasses for each platform and
provides a uniform interface for `H5vccAccessibilityImpl` to ask for the
TTS state in the platform and for the platform to communicate that said
state has not changed. While here, the Android code was converted to use
`base::ScopedObservation` instead of manual calls to `AddObserver()` and
`RemoveObserver()`.

This patch contains no changes in functionality.

Bug: 447135715