3p: Update locale tests and enable them on modular (#8478)
Introduce ScopedLocale to robustly manage locale_t resources in POSIX
locale NPLB tests. This ensures freelocale() is always invoked, even
on early exits, preventing resource leaks and flakiness. The
refactoring of the tests was done by Gemini, and then edited for
correctness.

This PR also enables the locale tests for modular builds. This makes
the modular builds now match the evergreen builds in regards to the
enabled locale tests.

Bug: 470106339