media: Support synchronous MSE init segment append
Introduce a new H5vcc setting, Media.AppendFirstSegmentSynchronously,
to allow the first media segment appended to a SourceBuffer to be
processed synchronously.

** There are no changes to the existing logic if this setting isn't
enabled. **

Performing the initial segment parsing and appending synchronously
reduces delays associated with asynchronous processing. This improves
media startup times for Cobalt applications.

This change doesn't conform to the Media Source Extension spec, where
the append is required to be asynchronous.  It also changes subtle
states inside media pipeline that affect its event sequence, which
could affect the web app if the latter depends on the sequence of event
firing.  Thus the change will be experimented in production to rule out
such issues.

Bug: 271167555